+++ src/enums/sets.ts
@@ -6,15 +6,40 @@
 export enum SetComparisonEnum {
   EQUALS = 'equals',
   NOT_EQUALS = 'not_equals',
+  SAME_ELEMENTS = 'same_elements',
 }
 
 export type SetComparisonOper = SetComparisonEnum | `${SetComparisonEnum}`;
+
+export enum SetIntersectionEnum {
+  DISJOINT = 'disjoint',
+  INTERSECTS = 'intersects',
+}
+
+export type SetIntersectionOper = SetIntersectionEnum | `${SetIntersectionEnum}`;
+
+export enum SetMembershipEnum {
+  INCLUDES = 'includes',
+  EXCLUDES = 'excludes',
+}
+
+export type SetMembershipOper = SetMembershipEnum | `${SetMembershipEnum}`;
+
+export enum SetArrayMembershipEnum {
+  CONTAINS_ALL = 'contains_all',
+  CONTAINS_ANY = 'contains_any',
+  EXCLUDES_ALL = 'excludes_all',
+}
+
+export type SetArrayMembershipOper = SetArrayMembershipEnum | `${SetArrayMembershipEnum}`;
 
 export enum SetRelationEnum {
   DISJOINT = 'disjoint',
   INTERSECTS = 'intersects',
   SUBSET_OF = 'subset_of',
   SUPERSET_OF = 'superset_of',
+  STRICT_SUBSET_OF = 'strict_subset_of',
+  STRICT_SUPERSET_OF = 'strict_superset_of',
 }
 
 export type SetRelationOper = SetRelationEnum | `${SetRelationEnum}`;
@@ -32,6 +57,8 @@ export type SetSizeOper = SetSizeEnum | `${SetSizeEnum}`;
 export enum SetStateEnum {
   IS_EMPTY = 'is_empty',
   IS_NOT_EMPTY = 'is_not_empty',
+  HAS_PRIMITIVES = 'has_primitives',
+  HAS_OBJECTS = 'has_objects',
 }
 
 export type SetStateOper = SetStateEnum | `${SetStateEnum}`;