+++ src/predicates/sets/index.ts
@@ -4,6 +4,9 @@
  * @module predicates/sets
  */
 export * from './setComparison.js';
+export * from './setIntersection.js';
+export * from './setMembership.js';
+export * from './setArrayMembership.js';
 export * from './setRelation.js';
 export * from './setSize.js';
 export * from './setState.js';